Hero is constrained to the map size
